:truck: Move some the client certificate screens around
