+++ src/rogallo/widgets/__init__.py
@@ -3,6 +3,7 @@
 ##############################################################################
 # Local imports.
 from .command_line import CommandLine
+from .data import Data
 from .side_panel import SidePanel
 from .toolbar import Toolbar
 from .viewer import Viewer
@@ -11,6 +12,7 @@ from .viewer import Viewer
 # Exports.
 __all__ = [
     "CommandLine",
+    "Data",
     "SidePanel",
     "Toolbar",
     "Viewer",
