+++ src/rogallo/widgets/side_panel/client_certificates.py
@@ -27,8 +27,10 @@ from ...data import Bookmarks, LocationHistory, NavigationHistory
 from ...document import Document
 from ...messages import ClientCertificatesModified
 from ...safe_escape import escape
-from ...screens.certificate_maker import ClientCertificateMaker
-from ...screens.certificate_viewer import ClientCertificateViewer
+from ...screens.client_certificate import (
+    ClientCertificateMaker,
+    ClientCertificateViewer,
+)
 from ...screens.scope_picker import ScopePicker
